bot/requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
pydantic-core==2.33.2
python-dotenv==1.0.0